+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed "copy text to clipboard" feature to use a fallback in case of `clipboard-write` permission is not granted (#90)
+- Reduced init time of `Widget` by lazy init for the report page views, i.e. init views only when the report page is opened for the first time
 
 ## 1.0.0-beta.64 (15-02-2022)