+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - Fixed "copy text to clipboard" feature to use a fallback in case of `clipboard-write` permission is not granted (#90)
 - Reduced init time of `Widget` by lazy init for the report page views, i.e. init views only when the report page is opened for the first time
-- Tweaked source view styles
+- Tweaked `source` view styles
+- Fixed badge views styles to change a color on hover only when `href` attribute is set
 
 ## 1.0.0-beta.64 (15-02-2022)