+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed exception when `lookup()` method of object marker is called with an unknown type, return `undefined` instead
 - Fixed hanging up when a pattern matches zero-length substring (i.e. in `input` view with type `regexp`)
 - Fixed edge cases in displaying an error location in query editor
+- Changed suggestions behaviour in query editor to be less intrusive
 
 ## 1.0.0-beta.64 (15-02-2022)