+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed edge cases in displaying an error location in query editor
 - Changed suggestions behaviour in query editor to be less intrusive
 - Changed style for suggestions popup, use icons for value types instead of captions
+- Fixed report page header editing to not re-render a report on input
 
 ## 1.0.0-beta.64 (15-02-2022)