+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Fixed exception when `lookup()` method of object marker is called with an unknown type, return `undefined` instead
 - Fixed hanging up when a pattern matches zero-length substring (i.e. in `input` view with type `regexp`)
 - Fixed edge cases in displaying an error location in query editor
+- Bumped jora to [1.0.0-beta.6](https://github.com/discoveryjs/jora/releases/tag/v1.0.0-beta.6)
+- Improved performance of variants generation for suggestions popup up to 30-40 times
 - Changed suggestions behaviour in query editor to be less intrusive
 - Changed style for suggestions popup, use icons for value types instead of captions
 - Fixed report page header editing to not re-render a report on input